DefaultAddressParser: move from o.b.core to base
(The remaining dependencies on `core` are on params p2sh header and
 in deprecated methods)
