+++ core/src/main/java/org/bitcoinj/base/LegacyAddress.java
@@ -21,7 +21,6 @@ package org.bitcoinj.base;
 import com.google.common.primitives.UnsignedBytes;
 import org.bitcoinj.base.exceptions.AddressFormatException;
 import org.bitcoinj.core.AddressMessage;
-import org.bitcoinj.core.DefaultAddressParser;
 import org.bitcoinj.core.ECKey;
 import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.core.PeerAddress;
