+++ core/src/test/java/org/bitcoinj/utils/MessageVerifyUtilsTest.java
@@ -1,7 +1,7 @@
 package org.bitcoinj.utils;
 
 import org.bitcoinj.base.AddressParser;
-import org.bitcoinj.core.DefaultAddressParser;
+import org.bitcoinj.base.DefaultAddressParser;
 import org.bitcoinj.core.NetworkParameters;
 import org.bitcoinj.params.MainNetParams;
 import org.bitcoinj.params.TestNet3Params;
